tests/test_processor.py
"""Test suite for processor."""

import os
import pytest
import numpy as np
from pathlib import Path
from unittest.mock import MagicMock, patch

from src.utils.core.processor import Processor


@pytest.fixture
def processor() -> Processor:
    """Create processor instance."""
    proc = Processor()
    return proc


def test_initialization(processor: Processor) -> None:
    """Test processor initialization."""
    assert processor.initialized
    assert processor.config is not None
    assert 'batch_size' in processor.config


def test_batch_processing(processor: Processor) -> None:
    """Test batch processing."""
    # Create test batch
    batch = [np.zeros((100, 100, 3), dtype=np.uint8) for _ in range(3)]
    
    # Process batch
    results = processor._process_batch(batch)
    assert len(results) == 3
    assert all(isinstance(r, np.ndarray) for r in results)


def test_error_handling(processor: Processor) -> None:
    """Test error handling."""
    # Test invalid path
    with pytest.raises(FileNotFoundError):
        processor.process_dataset("invalid_path", "test_session")
    
    # Test missing session ID
    with pytest.raises(ValueError):
        processor.process_dataset(".", "")


def test_resource_management(processor: Processor) -> None:
    """Test resource management."""
    # Mock resource check to simulate resource exhaustion
    with patch.object(processor.device_manager, 'check_resource_availability', return_value=False):
        with pytest.raises(RuntimeError, match="Resource limits exceeded"):
            processor.process_dataset("test_path", "test_session")


def test_session_management(processor: Processor) -> None:
    """Test session management."""
    session_id = "test_session"
    
    # Create session
    assert processor.session_manager.create_session(session_id, {})
    
    # Test duplicate session
    assert not processor.session_manager.create_session(session_id, {})
    
    # Cleanup session
    processor.session_manager.cleanup_session(session_id)


def test_system_monitoring(processor: Processor) -> None:
    """Test system monitoring."""
    # Get error info
    error_info = processor.get_error_info()
    assert isinstance(error_info, dict)
    assert 'error' in error_info
    assert 'timestamp' in error_info


def test_edge_cases(processor: Processor) -> None:
    """Test edge cases."""
    # Test empty batch
    results = processor._process_batch([])
    assert len(results) == 0
    
    # Test single item batch
    batch = [np.zeros((100, 100, 3), dtype=np.uint8)]
    results = processor._process_batch(batch)
    assert len(results) == 1
